Revert "Feature/proxies"
