+++ .env-linux
@@ -10,18 +10,3 @@ BROWSER=firefox
 DRIVER_HOST=localhost
 DRIVER_PORT=9000
 DRIVER_SESSION_COUNT=1
-AZURE_SEARCH_URI=https://sqeed-dev-bus.servicebus.windows.net/search/
-AZURE_SEARCH_DEQUEUE_API=https://sqeed-dev-bus.servicebus.windows.net/search/messages/head
-AZURE_PROFILE_URI=https://sqeed-dev-bus.servicebus.windows.net/profile/
-AZURE_PROFILE_DEQUEUE_API=https://sqeed-dev-bus.servicebus.windows.net/profile/messages/head
-AZURE_PROFILE_QUEUE_API=https://sqeed-dev-bus.servicebus.windows.net/profile/messages
-AZURE_MANAGER_BUS_URI=https://sqeed-dev-bus.servicebus.windows.net/manager/
-AZURE_MANAGER_BUS_API=https://sqeed-dev-bus.servicebus.windows.net/manager/messages
-AZURE_MANAGER_BUS_KEY=bC3swcT8ywbPHpNgSx4eJVG6tkhBtlC8b+ASbLzwa+4=
-AZURE_SAS_KEY_NAME=rw
-AZURE_SAS_PROFILE_KEY=xIf1mAf1YIRFq8WoUk4me4yG2XILTDUH7+ASbJl066Y=
-AZURE_SAS_SEARCH_KEY=kZxuQJeumvq2r7n+s7uhSENCDJIEdYjf6+ASbD/itM4=
-MANAGER_SEARCH_API=https://sqeed-ai-dev-backend-ca02.politepebble-0ee01f27.swedencentral.azurecontainerapps.io/crawler-manager/search-result
-MANAGER_PROFILE_API=https://sqeed-ai-dev-backend-ca02.politepebble-0ee01f27.swedencentral.azurecontainerapps.io/crawler-manager/parse-result
-LINKEDIN_USERNAME=kamil.psenak@cronus.sk
-LINKEDIN_PASSWORD=Fs2snloux2bEMpgQ9JSB
