+++ Cargo.toml
@@ -15,18 +15,8 @@ serde = "1.0.214"
 dotenvy = "0.15.7"
 fs_extra = "1.3.0"
 crossbeam = "0.8.4"
+serde_json = "1.0.132"
 async-scoped = { version = "0.9.0", features = ["use-tokio"] }
-reqwest = { version = "0.12.12", features = ["cookies", "multipart"] }
-reqwest_cookie_store = "0.8.0"
-hmac = { version = "0.12", features = ["std"] }
-sha2 = "0.10"
-base64 = "0.21"
-urlencoding = "2.1"
-serde_json = "1.0.135"
-cookie = "0.18.1"
-chrono = "0.4.39"
-[target.'cfg(target_os = "linux")'.dependencies]
-openssl = { version = "0.10", features = ["vendored"] }
 
 [lib]
 path = "src/lib.rs"
@@ -38,7 +28,3 @@ path = "src/main.rs"
 [[bin]]
 name = "omicron_crawler_server"
 path = "src/server.rs"
-
-[[bin]]
-name = "omicron_crawler_server_bus"
-path = "src/bus_server.rs"
