--- collections/azure-api/bruno.json
@@ -1,9 +1,0 @@
-{
-  "version": "1",
-  "name": "azure-api",
-  "type": "collection",
-  "ignore": [
-    "node_modules",
-    ".git"
-  ]
-}
