--- collections/azure-api/profile_put.bru
@@ -1,60 +1,0 @@
-meta {
-  name: profile_put
-  type: http
-  seq: 2
-}
-
-post {
-  url: https://sqeed-dev-bus.servicebus.windows.net/profile/messages
-  body: json
-  auth: none
-}
-
-headers {
-  ~api-version: 2022-01-01-preview
-}
-
-body:json {
-  {
-    "ids": [
-      "tomas-danis",
-      "ACoAACc9S0UBnls3lwgq156yD0kivVhIA98dtes",
-      "ACoAAAMD86wBp3VJ52OSSQ2bRCtKdadQLvmqzrU",
-      "ACoAACR-QR0B_M1s1sDm4yo6GNl56uRMQ3PHU1Y"
-    ]
-  }
-}
-
-body:text {
-  Hello World
-}
-
-body:xml {
-  Test1
-}
-
-script:pre-request {
-  const CryptoJS = require('crypto-js');
-  
-  function createServiceBusOrEventHubsSASToken(resourceUri, sasKeyName, sasKey) {
-      if (!resourceUri || !sasKeyName  || !sasKey) {
-          throw "Missing required parameter";
-      }
-      const encoded = encodeURIComponent(resourceUri);
-      const now = new Date();
-      const minute = 60;
-      const ttl = Math.round(now.getTime() / 1000) + minute;
-      const signature = encoded + '\n' + ttl;
-      const hash = CryptoJS.HmacSHA256(signature, sasKey).toString(CryptoJS.enc.Base64);
-      return 'SharedAccessSignature sr=' + encoded + '&sig=' +
-          encodeURIComponent(hash) + '&se=' + ttl + '&skn=' + sasKeyName;
-  }
-  
-  const uri = 'https://sqeed-dev-bus.servicebus.windows.net/profile/';  // Your Service Bus namespace
-  const sasKeyName = 'rw';  // The name of your Shared Access Policy
-  const sasKey = 'xIf1mAf1YIRFq8WoUk4me4yG2XILTDUH7+ASbJl066Y=';  // Your primary/secondary key
-  
-  const token = createServiceBusOrEventHubsSASToken(uri, sasKeyName, sasKey);
-  req.setHeader('Authorization', token);
-  req.setHeader('Date', new Date().toUTCString());
-}
