--- collections/linkedin_api/authenticate.bru
@@ -1,52 +1,0 @@
-meta {
-  name: authenticate
-  type: http
-  seq: 2
-}
-
-post {
-  url: {{LINKEDIN_URL}}/uas/authenticate
-  body: formUrlEncoded
-  auth: none
-}
-
-headers {
-  X-Li-User-Agent: LIAuthLibrary:0.0.3 com.linkedin.android:4.1.881 Asus_ASUS_Z01QD:android_9
-  User-Agent: ANDROID OS
-  X-User-Language: en
-  X-User-Locale: en_US
-  Accept-Language: en-us
-  csrf-token: {{JSESSIONID}}
-}
-
-body:json {
-  {
-    "session_key": "erik9631@gmail.com",
-    "session_password": "soRMoaN7C2bX2mKbV9V4",
-    "JSESSIONID": "{{JSESSIONID}}"
-  }
-}
-
-body:form-urlencoded {
-  session_key: kamil.psenak@cronus.sk
-  session_password: Fs2snloux2bEMpgQ9JSB
-  JSESSIONID: {{JSESSIONID}}
-}
-
-script:pre-request {
-  const cookies = bru.getEnvVar('cookies');
-  req.headers.Cookie = cookies;
-}
-
-script:post-response {
-  const cookies = res.headers.get("set-cookie");
-  console.log(cookies)
-  if(cookies === undefined){
-    console.log("No cookies found");
-    return;
-  }
-  //const session_entry = cookies[5];
-  // const JSESSIONID = session_entry.match(/JSESSIONID=(.*?);/)[1].replace(/"/g, '');
-  // bru.setEnvVar('JSESSIONID', JSESSIONID);
-  bru.setEnvVar('cookies', cookies);
-}
