--- collections/linkedin_api/bruno.json
@@ -1,9 +1,0 @@
-{
-  "version": "1",
-  "name": "linkedin_api",
-  "type": "collection",
-  "ignore": [
-    "node_modules",
-    ".git"
-  ]
-}
