--- collections/linkedin_api/challenge.bru
@@ -1,27 +1,0 @@
-meta {
-  name: challenge
-  type: http
-  seq: 9
-}
-
-post {
-  url: https://www.linkedin.com/checkpoint/lg/direct-login-submit?addDetailedLoginResult=add&dropLiatCookie=drop&csrfToken=ajax%3A4591103049031164980&pSIdString=56827d20-2fd0-4f7e-9362-93e65df36734&ut=1MDL2ONKGFVXA1
-  body: json
-  auth: none
-}
-
-params:query {
-  addDetailedLoginResult: add
-  dropLiatCookie: drop
-  csrfToken: ajax:4591103049031164980
-  pSIdString: 56827d20-2fd0-4f7e-9362-93e65df36734
-  ut: 1MDL2ONKGFVXA1
-}
-
-headers {
-  X-Li-User-Agent: LIAuthLibrary:0.0.3 com.linkedin.android:4.1.881 Asus_ASUS_Z01QD:android_9
-  User-Agent: ANDROID OS
-  X-User-Language: en
-  X-User-Locale: en_US
-  Accept-Language: en-us
-}
