--- collections/linkedin_api/environments/linkedin.bru
@@ -1,9 +1,0 @@
-vars {
-  LINKEDIN_URL: https://www.linkedin.com
-  API_URL: /voyager/api
-  cookies: bcookie="v=2&9579ad28-3572-4086-8d3c-b82851e4d584"; bscookie="v=1&2024111311410930621b54-42b5-467f-86d0-3c6d018b597fAQFx9NJNd9EMDa2fiTYceVw07WPzvkdW"; li_gc=MTsyMTsxNzM3NTc0MDQ0OzI7MDIxFUKgoBv8N1uvsN1uLHT6jRm3XQZg1gZPpzQmsxsQ0Y0=; li_mc=MTsyMTsxNzM3NjQwNzcwOzI7MDIx3IhyNMxWQtkGKrsgHZ5z+cxUdzIHWQ4Ht+XxEC/b8uk=; dfpfpt=b2febfaf26114608b937d81b73148c5d; timezone=Europe/Bratislava; li_theme=light; li_theme_set=app; UserMatchHistory=AQJavSPsW2y7nQAAAZSTfG9V1n60DY7kDWG_lCmYqOiN13cKBMnl1zPqH-6MDqrqYqgiiWQ1r4uQEq5Y8MXYXURgURPR6D4RlTVQtrzIUc19osqCJ5hjg9CEfyD0BbakTkGxh_jB6nuQepItMXNydPGUrqrgL5sBegDsS11qUc20_BAXjZvxgBTp2f1Ha89jxu9mkUDtm7bGk5P2Eejh5tG_QWi-je3J-gB4WbAqdeT_nMRWqBE6JjpbaAugU8rVCzCO5CJuUhk9kY84nN8Pvg7Ka8NhdaCnaZaSjiBFUmDcDlEucDJyzXbXUkxKtf4y4pz16tf7NJEjP2J3WUbh; li_rm=AQEXmFq6BbC9uwAAAZRXE6DrObQrqQkUVVtaVHzUAPJndwwVbga5wHXaWbrOrFSiPXyuQL2s_0HINxK8S3RtnsvnyWMk-T2NbBgpTnuHbVvWBSJHNo5vsBIeLaxO-f6iLd1xGDqeWks1RQjux8NWtlctCD0QKFSpytxST2Z0_2ahTUKqCbYTM1hT1JYHn7NABw6i2ee4MqhqfYAm7ughTKbF5maOwAOmFdpzL3Ngg2DrETy07ljVKloaFf2Rl5r2h1eWj-L2Z-IYlhnZ98VMY8VrCK_AplYxNEV5R1UKe-fRM4K-nyB_C8TkrUp2sPYEgy9HTobBValmwzaMrkA; visit=v=1&M; JSESSIONID="ajax:2321504918253678118"; li_alerts=e30=; g_state={"i_l":0}; li_at=AQEDAUt7mGIFD-b9AAABlI_laA8AAAGUs_HsD00ASZ1MoUVudbfs5esoVzE6K0ANO_gOAtAbRa_V4g1C9JKCBeNTMpFsxfwFau37jjQSm50upH5D_eHKv4I755jcoSUoMTbEpMxSCgpUi8HZ0FblQ1gd; liap=true; lang=v=2&lang=en-us; lidc="b=VB38:s=V:r=V:a=V:p=V:g=122:u=182:x=1:i=1737640147:t=1737644250:v=2:sig=AQGoOUcyPjHVC04IRgnXd4BmOH-v3voa"; __cf_bm=45pG6PTEa_FK2UlzLgiVUNe8.iL1b_qOzMTz7eJiNR8-1737641190-1.0.1.1-dmCxLibLRvoibMp0wdjqIWkHeF1sWbanOYpaVTpgkXAYkEQcCzlSyufZ.JBXm5NBgcshZJ_icP8wyHSiiRyX2A; fptctx2=taBcrIH61PuCVH7eNCyH0B9zcK90d%252bIeoo1r5v7Zc26LHEDTxXT2QMJYrFWf8XQke09KCiEXvtXCQA%252fThSVVQU3bVW%252bZfKMBFLLdjFBICmloUZINzsohc0QzrFZjpjqtVlgaIlNe0HrCrXpZe3JXZWGParo5aXqset9IQWd9e%252bBmLCNhqxRcqDUYOwp%252b3gdQfC9BCRqGFHjfyiUZsGG8O8ZHLhDFNkfM9hEktdiiD7BK2VptxNavzJGaCzq9ovWw9k5y6m7n6dLkc5mPtT8SuL9SDiDc04aHJlpLQWs9Eair3mjMRehU7vtumdP2zayAHyKr0crdOjQYDtwqR5grRIr4ES38hjTqyVxSbmtiyno%253d
-  JSESSIONID: ajax:2321504918253678118
-  PAGE: 2
-  KEYWORDS: Java
-  URN_ID: ACoAAAWs1dABCvfhO0Y2aqYxDtCEtuFntb6jfb4
-}
