--- collections/linkedin_api/people_search.bru
@@ -1,36 +1,0 @@
-meta {
-  name: people_search
-  type: http
-  seq: 6
-}
-
-get {
-  url: {{LINKEDIN_URL}}{{API_URL}}/graphql?variables=(start:{{PAGE}},origin:GLOBAL_SEARCH_HEADER,query:(keywords:{{KEYWORDS}},flagshipSearchIntent:SEARCH_SRP,queryParameters:List((key:resultType,value:List(PEOPLE))),includeFiltersInResponse:false))&queryId=voyagerSearchDashClusters.b0928897b71bd00a5a7291755dcd64f0
-  body: none
-  auth: none
-}
-
-params:query {
-  variables: (start:{{PAGE}},origin:GLOBAL_SEARCH_HEADER,query:(keywords:{{KEYWORDS}},flagshipSearchIntent:SEARCH_SRP,queryParameters:List((key:resultType,value:List(PEOPLE))),includeFiltersInResponse:false))
-  queryId: voyagerSearchDashClusters.b0928897b71bd00a5a7291755dcd64f0
-}
-
-headers {
-  X-Li-User-Agent: LIAuthLibrary:0.0.3 com.linkedin.android:4.1.881 Asus_ASUS_Z01QD:android_9
-  User-Agent: ANDROID OS
-  X-User-Language: en
-  X-User-Locale: en_US
-  Accept-Language: en-us
-  csrf-token: {{JSESSIONID}}
-  x-restli-protocol-version: 2.0.0
-}
-
-script:pre-request {
-  const cookies = bru.getEnvVar('cookies');
-  req.headers.Cookie = cookies;
-}
-
-script:post-response {
-  const body = res.body;
-  console.log(body.data.searchDashClustersByAll.elements[0].items.length)
-}
