--- collections/linkedin_api/profile_contact_info.bru
@@ -1,26 +1,0 @@
-meta {
-  name: profile_contact_info
-  type: http
-  seq: 5
-}
-
-get {
-  url: {{LINKEDIN_URL}}{{API_URL}}/identity/profiles/matus-chochlik-154a7827/profileContactInfo
-  body: none
-  auth: none
-}
-
-headers {
-  X-Li-User-Agent: LIAuthLibrary:0.0.3 com.linkedin.android:4.1.881 Asus_ASUS_Z01QD:android_9
-  User-Agent: ANDROID OS
-  X-User-Language: en
-  X-User-Locale: en_US
-  Accept-Language: en-us
-  csrf-token: {{JSESSIONID}}
-  x-restli-protocol-version: 2.0.0
-}
-
-script:pre-request {
-  const cookies = bru.getEnvVar('cookies');
-  req.headers.Cookie = cookies;
-}
