--- collections/linkedin_api/profile_experience.bru
@@ -1,38 +1,0 @@
-meta {
-  name: profile_experience
-  type: http
-  seq: 8
-}
-
-get {
-  url: {{LINKEDIN_URL}}{{API_URL}}/graphql?variables=(start:1,profileUrn:urn%3Ali%3Afsd_profile%3A{{URN_ID}},sectionType:experience)&queryId=voyagerIdentityDashProfileComponents.7af5d6f176f11583b382e37e5639e69e&includeWebMetadata=true
-  body: none
-  auth: none
-}
-
-params:query {
-  variables: (start:1,profileUrn:urn:li:fsd_profile:{{URN_ID}},sectionType:experience)
-  queryId: voyagerIdentityDashProfileComponents.7af5d6f176f11583b382e37e5639e69e
-  includeWebMetadata: true
-}
-
-headers {
-  X-Li-User-Agent: LIAuthLibrary:0.0.3 com.linkedin.android:4.1.881 Asus_ASUS_Z01QD:android_9
-  User-Agent: ANDROID OS
-  X-User-Language: en
-  X-User-Locale: en_US
-  Accept-Language: en-us
-  csrf-token: {{JSESSIONID}}
-  x-restli-protocol-version: 2.0.0
-  accept: application/vnd.linkedin.normalized+json+2.1
-}
-
-script:pre-request {
-  const cookies = bru.getEnvVar('cookies');
-  req.headers.Cookie = cookies;
-}
-
-script:post-response {
-  const body = res.body;
-  console.log(body)
-}
