--- collections/linkedin_api/profile_skills.bru
@@ -1,31 +1,0 @@
-meta {
-  name: profile_skills
-  type: http
-  seq: 6
-}
-
-get {
-  url: {{LINKEDIN_URL}}{{API_URL}}/identity/profiles/lukáš-oščipovský-612326aa/skills?start=0&count=100
-  body: none
-  auth: none
-}
-
-params:query {
-  start: 0
-  count: 100
-}
-
-headers {
-  X-Li-User-Agent: LIAuthLibrary:0.0.3 com.linkedin.android:4.1.881 Asus_ASUS_Z01QD:android_9
-  User-Agent: ANDROID OS
-  X-User-Language: en
-  X-User-Locale: en_US
-  Accept-Language: en-us
-  csrf-token: {{JSESSIONID}}
-  x-restli-protocol-version: 2.0.0
-}
-
-script:pre-request {
-  const cookies = bru.getEnvVar('cookies');
-  req.headers.Cookie = cookies;
-}
