--- collections/linkedin_api/session_cookies.bru
@@ -1,27 +1,0 @@
-meta {
-  name: session_cookies
-  type: http
-  seq: 1
-}
-
-get {
-  url: {{LINKEDIN_URL}}/uas/authenticate
-  body: json
-  auth: none
-}
-
-headers {
-  X-Li-User-Agent: LIAuthLibrary:0.0.3 com.linkedin.android:4.1.881 Asus_ASUS_Z01QD:android_9
-  User-Agent: ANDROID OS
-  X-User-Language: en
-  X-User-Locale: en_US
-  Accept-Language: en-us
-}
-
-script:post-response {
-  const cookies = res.headers.get("set-cookie");
-  const session_entry = cookies[1];
-  const JSESSIONID = session_entry.match(/JSESSIONID=(.*?);/)[1].replace(/"/g, '');
-  bru.setEnvVar('JSESSIONID', JSESSIONID);
-  bru.setEnvVar('cookies', cookies);
-}
