--- collections/omicron/environments/omicron.bru
@@ -1,2 +1,0 @@
-vars {
-}
