+++ collections/omicron/profile.bru
@@ -5,15 +5,87 @@ meta {
 }
 
 post {
-  url: 127.0.0.1:8080/profiles
+  url: 74.241.242.176:8080/profiles
   body: json
   auth: none
 }
 
 body:json {
-  {
-    "ids": [
-      "tomas-danis"
-    ]
-  }
+  [
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAAAWs1dABZXg7RDqKugFxlSeo7gasFL1FPHQ,NAME_SEARCH,cypw?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAAAy0mZcBlmERvP-yDTL3gnlTLSELF6c7hrk,NAME_SEARCH,UDAQ?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAAAcDK_UBYZxeRVNrxjxmgsHvIHhg3xby2Lc,NAME_SEARCH,7vTn?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAABrBNPsB4MPIvpE3HbWVxETTQJ8TOIKp-04,NAME_SEARCH,FReR?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAAC7xxs0BVWmgs6nH77qWXRQwLie6kNxmmQQ,NAME_SEARCH,S7g-?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAAAp2oSIBAZW3aglP_zfF_XgprTWLVONsO-s,NAME_SEARCH,2NaR?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAACnEZccBuFomV2Q-kWcny2DKU4ZYJ1cahg8,NAME_SEARCH,3Gfu?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAAEcky30BfARqO1NX9OJ28AubC7DA7QLHA0w,NAME_SEARCH,iJ8H?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAAA9rud0B81Ces1k2ws0jr_rHusTENsMFGJU,NAME_SEARCH,ftr6?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAADPokucBNDkQN2WoiI3USCGEhDHGtjmH9I8,NAME_SEARCH,h167?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAAAlJO80BeP_D1gOoQgJJ30M36sk47x5tW5U,NAME_SEARCH,ImgT?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAABG1suwB-S2dvD44cijG-SbfJFJtvoqbRZ4,NAME_SEARCH,55Bo?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAABP7nh4BTRx-Ygy8DukhXNqJTWKz4kqqNGg,NAME_SEARCH,69yw?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAACHU0rMBJo9IVsYGhvuxGkar9-B-hpUbqDQ,NAME_SEARCH,O28m?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAADRpaagB1KE6kGGVK40QeRKPsA6K9X3PnD4,NAME_SEARCH,_C_M?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAAD0GYV0BAzpqwvw9GQW3124E40dXgNAwXag,NAME_SEARCH,Rhbq?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAADChduUBiEOBW0dCrCEIM_phZ2PNgRmTUP0,NAME_SEARCH,7MAM?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAACCthgIBBPwGH2m400WjAXyfQzrqBgt-Daw,NAME_SEARCH,DYUO?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAABPPg1oBURMxdqUWVvzT4oEG7gyUtsxM5ew,NAME_SEARCH,t3kJ?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAABgi9EMB_spy-5pPFc6qsbF2rMjRkfBAoEA,NAME_SEARCH,mw3i?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAAAVTKhkBj-83IEUoar9FbF-uLoA0BJxM9n8,NAME_SEARCH,zvkC?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAAADiJnQBi62QSVRz8ONyTN8j4RMp_qUBSeE,NAME_SEARCH,FZna?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAABePdVABvql4r-z00BTlrySgpK77y1Adf0o,NAME_SEARCH,WLIc?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAAArgC9UBpRSK4ZpQrkVpCWkOS5AGPHILITw,NAME_SEARCH,14Et?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    },
+    {
+      "sales_url": "https://www.linkedin.com/sales/lead/ACwAAAuMYKgBkDx12ooPP0Ceh8X2Eos1PORzuYc,NAME_SEARCH,QCyE?_ntb=KAHhkeRAT0ejheHlZJlIrg%3D%3D"
+    }
+  ]
 }
