+++ collections/omicron/search.bru
@@ -12,8 +12,7 @@ post {
 
 body:json {
   {
-    "keywords":"Rust",
-    "page": 0,
-    "end": 10
+    "keywords": "Java",
+    "geography": "Slovakia"
   }
 }
