+++ src/lib.rs
@@ -2,7 +2,6 @@
 extern crate log;
 #[macro_use]
 pub mod macros;
-pub mod azure;
 pub mod driver;
 pub mod env;
 pub mod errors;
