+++ src/linkedin/mod.rs
@@ -1,2 +1,4 @@
-pub mod api;
-pub mod web_driver;
+pub mod crawler;
+pub mod enums;
+pub mod parse_sales;
+pub mod profiles;
