--- src/linkedin/web_driver/mod.rs
@@ -1,6 +1,0 @@
-pub mod crawler;
-pub mod enums;
-pub mod linkedin;
-pub mod profiles;
-pub mod sales;
-pub mod sales_crawler;
