+++ src/logger.rs
@@ -1,4 +1,3 @@
-use chrono::Utc;
 use log::{LevelFilter, Metadata, Record};
 use std::sync::Once;
 
@@ -12,8 +11,7 @@ impl log::Log for Logger {
     }
 
     fn log(&self, record: &Record) {
-        let now = Utc::now();
-        println!("[{}][{}] {}", now.to_rfc3339(), record.level(), record.args());
+        println!("[{}] {}", record.level(), record.args());
     }
 
     fn flush(&self) {}
