+++ src/main.rs
@@ -2,8 +2,10 @@
 extern crate log;
 
 use log::LevelFilter;
-use omicron_crawler::azure::AzureClient;
-use omicron_crawler::env::load_env;
+use omicron_crawler::driver::service::GeckoDriverService;
+use omicron_crawler::driver::session_manager::SessionManager;
+use omicron_crawler::env::{get_env, load_env};
+use omicron_crawler::linkedin::crawler::Crawler;
 use omicron_crawler::logger::Logger;
 
 //TODO
@@ -14,33 +16,20 @@ use omicron_crawler::logger::Logger;
 async fn main() {
     Logger::init(LevelFilter::Trace);
     load_env();
-    let azure_client = AzureClient::new().await;
-    let response = match azure_client.dequeue_search().await {
-        Ok(response) => response,
-        Err(e) => panic!("Failed to dequeue profile {}", e),
-    };
-    // load_env();
-    //
-    // let env = get_env().await;
-    // let manager: SessionManager<GeckoDriverService> = SessionManager::new(
-    //     env.driver_host.as_str(),
-    //     env.driver_port,
-    //     env.driver_session_count,
-    //     env.driver_path.as_str(),
-    //     env.profile_path.as_str(),
-    //     env.browser_binary_path.as_deref(),
-    // )
-    // .await;
-    // let pool = &manager.pool;
-    // let session = pool.acquire().unwrap();
-    // let crawler = LinkedinCrawler::new(session).await;
-    //
-    // match crawler.parse_profile("https://www.linkedin.com/in/matus-chochlik-154a7827/").await {
-    //     Ok(profile) => println!("{}", profile),
-    //     Err(e) => println!("{}", e),
-    // }
-    //
-    // tokio::time::sleep(Duration::from_secs(5)).await;
+
+    let env = get_env().await;
+    let manager: SessionManager<GeckoDriverService> = SessionManager::new(
+        env.driver_host.as_str(),
+        env.driver_port,
+        env.driver_session_count,
+        env.driver_path.as_str(),
+        env.profile_path.as_str(),
+        env.browser_binary_path.as_deref(),
+    )
+    .await;
+    let pool = &manager.pool;
+    let session = pool.acquire().unwrap();
+    let crawler = Crawler::new(session).await;
 
     // let result = crawler
     //     .parse_profile("https://www.linkedin.com/sales/lead/ACwAABtsCsMBHH4i-dKLOpQqrcqE4H3YDX8CbxE")
