+++ src/server.rs
@@ -1,32 +1,41 @@
 pub mod services;
 
 use crate::services::hello;
-use crate::services::request::{profiles, search};
+use crate::services::request::{message, profiles, search};
 use actix_web::{App, HttpServer};
+use omicron_crawler::driver::service::GeckoDriverService;
+use omicron_crawler::driver::session_manager::SessionManager;
 use omicron_crawler::env::{get_env, load_env};
-use omicron_crawler::linkedin::api::crawler::Crawler;
-use omicron_crawler::linkedin::api::rate_limits::RateLimiter;
 use omicron_crawler::logger::Logger;
 use tokio::sync::OnceCell;
-static CRAWLER: OnceCell<Crawler> = OnceCell::const_new();
 
-pub async fn get_crawler() -> &'static Crawler {
-    CRAWLER
+static DRIVER_SESSION_MANAGER: OnceCell<SessionManager<GeckoDriverService>> = OnceCell::const_new();
+
+pub async fn get_driver_session_manager() -> &'static SessionManager<GeckoDriverService> {
+    DRIVER_SESSION_MANAGER
         .get_or_init(|| async {
-            let username = get_env().await.linkedin_username.as_str();
-            let password = get_env().await.linkedin_password.as_str();
-            let mut crawler = Crawler::new(RateLimiter::new(100, 800), username, password).await;
-            crawler
+            let env = get_env().await;
+            SessionManager::<GeckoDriverService>::new(
+                env.driver_host.as_str(),
+                env.driver_port,
+                env.driver_session_count,
+                env.driver_path.as_str(),
+                env.profile_path.as_str(),
+                env.browser_binary_path.as_deref(),
+            )
+            .await
         })
         .await
 }
+
 #[tokio::main(flavor = "multi_thread")]
 async fn main() -> std::io::Result<()> {
     load_env();
     let env = get_env().await;
     Logger::init(env.log_level);
-    get_crawler().await;
-    let result = HttpServer::new(|| App::new().service(hello).service(search).service(profiles))
+    get_driver_session_manager().await;
+
+    let result = HttpServer::new(|| App::new().service(hello).service(search).service(profiles).service(message))
         .bind((env.host.as_str(), env.port))?
         .system_exit()
         .run()
