+++ tests/mod.rs
@@ -6,6 +6,3 @@ extern crate omicron_crawler;
 mod chrome_driver_tests;
 #[cfg(test)]
 mod chrome_parsing_tests;
-
-#[cfg(test)]
-mod api_tests;
